+++ sentry_streams/tests/adapters/arroyo/test_broadcaster.py
@@ -10,7 +10,7 @@ from sentry_kafka_schemas.schema_types.ingest_metrics_v1 import IngestMetric
 from sentry_streams.adapters.arroyo.broadcaster import Broadcaster
 from sentry_streams.adapters.arroyo.routes import Route
 from sentry_streams.pipeline.message import PyMessage
-from tests.adapters.arroyo.message_helpers import make_value_msg
+from tests.adapters.arroyo.helpers.message_helpers import make_value_msg
 
 
 def test_submit_routedvalue(metric: IngestMetric) -> None:
