+++ sentry_streams/tests/adapters/arroyo/test_forwarder.py
@@ -13,7 +13,7 @@ from sentry_kafka_schemas.schema_types.ingest_metrics_v1 import IngestMetric
 from sentry_streams.adapters.arroyo.forwarder import Forwarder
 from sentry_streams.adapters.arroyo.routes import Route, RoutedValue
 from sentry_streams.pipeline.message import PyMessage as StreamsMessage
-from tests.adapters.arroyo.message_helpers import make_msg
+from tests.adapters.arroyo.helpers.message_helpers import make_msg
 
 
 def test_submit(metric: IngestMetric) -> None:
