+++ sentry_streams/tests/adapters/arroyo/test_reduce_delegate.py
@@ -1,5 +1,5 @@
 from datetime import datetime
-from typing import MutableSequence, Optional, Sequence, Tuple, TypeVar, Union
+from typing import MutableSequence, Optional, Sequence, TypeVar, Union
 
 from arroyo.processing.strategies.abstract import ProcessingStrategy
 from arroyo.types import FilteredPayload
@@ -14,17 +14,16 @@ from sentry_streams.adapters.arroyo.reduce_delegate import (
 from sentry_streams.adapters.arroyo.routes import RoutedValue
 from sentry_streams.adapters.arroyo.rust_step import (
     ArroyoStrategyDelegate,
-    Committable,
     OutputRetriever,
 )
 from sentry_streams.pipeline.message import (
     Message,
     PyMessage,
-    RustMessage,
     rust_msg_equals,
 )
 from sentry_streams.pipeline.pipeline import Batch
-from tests.adapters.arroyo.message_helpers import build_py_msg, build_rust_msg
+from tests.adapters.arroyo.helpers.delegate_helpers import assert_equal_batches
+from tests.adapters.arroyo.helpers.message_helpers import build_py_msg, build_rust_msg
 
 TStrategyOut = TypeVar("TStrategyOut")
 
@@ -106,17 +105,6 @@ class FakeReducer(ProcessingStrategy[Union[FilteredPayload, RoutedValue]]):
 
     def terminate(self) -> None:
         self.__next.terminate()
-
-
-def assert_equal_batches(
-    batch1: Sequence[Tuple[RustMessage, Committable]],
-    batch2: Sequence[Tuple[RustMessage, Committable]],
-) -> None:
-    assert len(batch1) == len(batch2)
-    for i, msg1 in enumerate(batch1):
-        msg2 = batch2[i]
-        assert msg1[0].payload == msg2[0].payload, f"Payload mismatch at index {i}"
-        assert msg1[1] == msg2[1], f"Committable mismatch at index {i}"
 
 
 def test_reduce_poll() -> None:
